Create `keytypes` package
