+++ cmd/cometbft/commands/run_node.go
@@ -3,15 +3,12 @@ package commands
 import (
 	"encoding/hex"
 	"fmt"
-	"strings"
 
 	"github.com/spf13/cobra"
 
 	"github.com/cometbft/cometbft/crypto"
-	"github.com/cometbft/cometbft/crypto/bls12381"
 	"github.com/cometbft/cometbft/crypto/ed25519"
-	"github.com/cometbft/cometbft/crypto/secp256k1"
-	"github.com/cometbft/cometbft/crypto/sr25519"
+	kt "github.com/cometbft/cometbft/internal/keytypes"
 	cmtos "github.com/cometbft/cometbft/internal/os"
 	nm "github.com/cometbft/cometbft/node"
 )
@@ -19,49 +16,10 @@ import (
 var (
 	genesisHash []byte
 	keyType     string
-	keyTypes    map[string]func() (crypto.PrivKey, error)
 )
 
-func init() {
-	keyTypes = map[string]func() (crypto.PrivKey, error){
-		ed25519.KeyType: func() (crypto.PrivKey, error) { //nolint: unparam
-			return ed25519.GenPrivKey(), nil
-		},
-		secp256k1.KeyType: func() (crypto.PrivKey, error) { //nolint: unparam
-			return secp256k1.GenPrivKey(), nil
-		},
-		sr25519.KeyType: func() (crypto.PrivKey, error) { //nolint: unparam
-			return sr25519.GenPrivKey(), nil
-		},
-		bls12381.KeyType: func() (crypto.PrivKey, error) {
-			pk, err := bls12381.GenPrivKey()
-			if err != nil {
-				return nil, fmt.Errorf("failed to generate BLS key: %w", err)
-			}
-			return pk, nil
-		},
-	}
-}
-
 func genPrivKeyFromFlag() (crypto.PrivKey, error) {
-	return genPrivKey(keyType)
-}
-
-func genPrivKey(keyType string) (crypto.PrivKey, error) {
-	genF, ok := keyTypes[keyType]
-	if !ok {
-		return nil, fmt.Errorf("unsupported key type: %s", keyType)
-	}
-	return genF()
-}
-
-func listKeyTypes() string {
-	keyTypesS := make([]string, 0, len(keyTypes))
-	for k := range keyTypes {
-		keyTypesS = append(keyTypesS, fmt.Sprintf("%q", k))
-	}
-	keyTypes := strings.Join(keyTypesS, ", ")
-	return keyTypes
+	return kt.GenPrivKey(keyType)
 }
 
 // AddNodeFlags exposes some common configuration options on the command-line
@@ -132,7 +90,7 @@ func AddNodeFlags(cmd *cobra.Command) {
 		"db_dir",
 		config.DBPath,
 		"database directory")
-	cmd.Flags().StringVarP(&keyType, "key-type", "k", ed25519.KeyType, fmt.Sprintf("private key type (one of %s)", listKeyTypes()))
+	cmd.Flags().StringVarP(&keyType, "key-type", "k", ed25519.KeyType, fmt.Sprintf("private key type (one of %s)", kt.ListSupportedKeyTypes()))
 }
 
 // NewRunNodeCmd returns the command that allows the CLI to start a node.
